requirements.txt
google-api-python-client>=1.7.9
google-auth>=1.5.1
google-auth-httplib2>=0.0.3
google-auth-oauthlib>=0.3.0
python-dateutil>=2.8.0
discord.py>=1.1.1